--- assets/js/services/elasticsearch.test.js
@@ -1,5 +1,0 @@
-describe("elasticsearch.js service module", () => {
-  describe("Elasticsearch query parser", () => {
-    it("should do this and that", () => {});
-  });
-});
