+++ assets/js/services/reactive-search.js
@@ -167,3 +167,52 @@ export const FACET_SENSORS = [
 
 export const RESULT_SENSOR = "SearchResult";
 export const SEARCH_SENSOR = "SearchSensor";
+
+/**
+ * Grab the search string and any facet values from an ElasticSearch query object
+ * @param {Object} query An ElasticSearch query object
+ * @returns { Object } Map object with keys "search" and "terms"
+ */
+export function extractQueryParts(query = {}) {
+  const searchTermKey = "query_string";
+  const facetTermsKey = "terms";
+
+  let returnObj = {
+    search: "",
+    terms: {},
+  };
+
+  function findQueryParts(child) {
+    if (typeof child !== "object" && child !== null) {
+      return;
+    }
+
+    // Array value, loop through child objects
+    if (Array.isArray(child)) {
+      for (let arrayObj of child) {
+        findQueryParts(arrayObj);
+      }
+    }
+
+    // Grab search string
+    if (child.hasOwnProperty(searchTermKey)) {
+      returnObj.search = child[searchTermKey].query;
+      return;
+    }
+
+    // Grab facet terms
+    if (child.hasOwnProperty(facetTermsKey)) {
+      returnObj.terms = { ...returnObj.terms, ...child[facetTermsKey] };
+      return;
+    }
+
+    for (let property in child) {
+      findQueryParts(child[property]);
+    }
+  }
+
+  // Kick off recursive function
+  findQueryParts(query);
+
+  return returnObj;
+}
